fix: typescript styling errors
