+++ src/pages/dashboard/styled.ts
@@ -1,5 +1,7 @@
 import { styled } from '@mui/system';
 
-export const StyledDashboard = styled(`div`)(({ theme }) => ({
+const StyledDashboard = styled(`div`)(({ theme }) => ({
   backgroundColor: theme.palette.background.default,
 }));
+
+export default StyledDashboard;
